+++ components/src/main/java/com/mercadolibre/android/andesui/card/factory/AndesCardConfigurationFactory.kt
@@ -2,18 +2,22 @@ package com.mercadolibre.android.andesui.card.factory
 
 import android.content.Context
 import android.graphics.Typeface
+import android.view.View
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.card.bodyPadding.AndesCardBodyPadding
 import com.mercadolibre.android.andesui.card.hierarchy.AndesCardHierarchy
 import com.mercadolibre.android.andesui.card.padding.AndesCardPadding
 import com.mercadolibre.android.andesui.card.padding.AndesCardPaddingInterface
 import com.mercadolibre.android.andesui.card.style.AndesCardStyleInterface
+import com.mercadolibre.android.andesui.card.type.AndesCardType
 import com.mercadolibre.android.andesui.card.type.AndesCardTypeInterface
 import com.mercadolibre.android.andesui.color.AndesColor
 import com.mercadolibre.android.andesui.color.toAndesColor
 import com.mercadolibre.android.andesui.typeface.getFontOrDefault
 
 internal data class AndesCardConfiguration(
+    val title: String?,
+    val titleVisibility: Int,
     val titleSize: Float,
     val titleTypeface: Typeface,
     val titleColor: AndesColor,
@@ -22,7 +26,9 @@ internal data class AndesCardConfiguration(
     val bodyPadding: AndesCardBodyPadding,
     val elevation: Float,
     val pipeColor: AndesColor,
-    val linkColor: AndesColor
+    val linkColor: AndesColor,
+    val pipeVisibility: Int,
+    val groupLinkVisibility: Int
 )
 
 internal object AndesCardConfigurationFactory {
@@ -30,6 +36,8 @@ internal object AndesCardConfigurationFactory {
     fun create(context: Context, andesCardAttrs: AndesCardAttrs): AndesCardConfiguration {
         return with(andesCardAttrs) {
             AndesCardConfiguration(
+                    title = andesCardAttrs.andesCardTitle,
+                    titleVisibility = resolveTitleVisibility(andesCardAttrs.andesCardTitle),
                     titleSize = resolveTitleSize(context),
                     titleTypeface = resolveTitleTypeface(context),
                     titleColor = resolveTitleColor(),
@@ -40,7 +48,9 @@ internal object AndesCardConfigurationFactory {
                             context, andesCardAttrs.andesCardHierarchy, andesCardAttrs.andesCardStyle.style
                     ),
                     pipeColor = resolvePipeColor(andesCardAttrs.andesCardType.type),
-                    linkColor = resolveLinkColor()
+                    linkColor = resolveLinkColor(),
+                    pipeVisibility = resolvePipeVisibility(andesCardAttrs.andesCardType, andesCardAttrs.andesCardHierarchy),
+                    groupLinkVisibility = resolverGroupLinkVisibility(andesCardAttrs.linkText, andesCardAttrs.andesCardHierarchy)
             )
         }
     }
@@ -63,4 +73,24 @@ internal object AndesCardConfigurationFactory {
     ) = style.elevation(context, hierarchy)
     private fun resolvePipeColor(type: AndesCardTypeInterface) = type.primaryColor()
     private fun resolveLinkColor() = R.color.andes_accent_color_500.toAndesColor()
+    private fun resolvePipeVisibility(type: AndesCardType, hierarchy: AndesCardHierarchy): Int {
+        return if (hierarchy == AndesCardHierarchy.PRIMARY)
+            type.type.pipeVisibility()
+        else
+            View.GONE
+    }
+    private fun resolveTitleVisibility(title: String?): Int {
+        return if (title.isNullOrEmpty()) {
+            View.GONE
+        } else {
+            View.VISIBLE
+        }
+    }
+
+    private fun resolverGroupLinkVisibility(linkText: String?, hierarchy: AndesCardHierarchy): Int {
+        return if (linkText.isNullOrEmpty() || hierarchy != AndesCardHierarchy.PRIMARY) {
+            View.GONE
+        } else {
+            View.VISIBLE
+        }        }
 }